data: atualiza mappers e entidades para suporte a standings e pontuação
- Altera o tipo do campo `points` de `Int` para `Float` no modelo de domínio e na entidade `DriverEntity`.
- Cria o arquivo `DriverMapper.kt` para centralizar a lógica de conversão entre DTOs, entidades e objetos de domínio de pilotos.
- Atualiza o `TeamMapper` para realizar o mapeamento a partir de `ConstructorStandingDto`, refletindo a estrutura de classificação.
- Ajusta a `TeamEntity` e o modelo `Team` para focar em dados de performance como `position`, `points` e `wins`.
- Adiciona o método `getTeamById` ao `TeamDao` para permitir a recuperação de um time específico de forma assíncrona.
- Implementa a concatenação de nome e sobrenome de pilotos durante o mapeamento de DTO para domínio/entidade.

+++ app/src/main/java/io/github/bortoletoeric/f1teamapp/domain/model/Driver.kt
@@ -5,5 +5,5 @@ data class Driver(
     val teamId: String,
     val name: String,
     val photoUrl: String?,
-    val points: Int
+    val points: Float
 )
